code cleaning and add examples to unit test
