+++ src/mlpro_mpps/pool/comps/PS002_C002_Pump1.py
@@ -6,10 +6,11 @@
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
 ## -- 2023-01-19  0.0.0     ML       Creation
+## -- 2023-03-28  1.0.0     ML/SY    Release of first version
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.0.0 (2023-01-19)
+Ver. 1.0.0 (2023-03-28)
 
 This module provides a default implementation of a component of the Liquid Station, which is a Pump.
 This pump is located on the input side of the Liquid Station to fill liquid into the tank.
