+++ src/mlpro_mpps/pool/comps/PS002_C003_Pump2.py
@@ -6,10 +6,11 @@
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
 ## -- 2023-01-19  0.0.0     ML       Creation
+## -- 2023-03-28  1.0.0     ML/SY    Release of first version
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.0.0 (2023-01-19)
+Ver. 1.0.0 (2023-03-28)
 
 This module provides a default implementation of a component of the Liquid Station, which is a Pump.
 This is one of two pumps located on the output side of the Liquid Station to empty liquid out of the tank.
