+++ test/test_examples.py
@@ -7,7 +7,7 @@
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
 ## -- 2023-02-27  0.0.0     SY/MRD   Creation
 ## -- 2023-02-27  1.0.0     SY/MRD   Release First Version
-## -- 2023-03-28  1.0.1     SY/MRD   Replacing importlib with runpy
+## -- 2023-03-28  1.0.1     SY/MRD   Replacing importlib with runpy, add howtos
 ## -------------------------------------------------------------------------------------------------
 
 
@@ -30,6 +30,9 @@ howto_list = {
     "ht_002": "mlpro_mpps.examples.howto_002_set_up_MPPS",
     "ht_003": "mlpro_mpps.examples.howto_003_run_RL_on_BGLP_using_MPPS",
     "ht_004": "mlpro_mpps.examples.howto_004_run_GT_on_BGLP_using_MPPS",
+    "ht_005": "mlpro_mpps.examples.howto_005_run_RL_on_LS_using_MPPS",
+    "ht_006": "mlpro_mpps.examples.howto_006_run_EA_on_LS_using_MPPS",
+    "ht_007": "mlpro_mpps.examples.howto_007_run_using_LS_MPPS_data_generation",
 }
